Added support for "stub" objects created from imported JSON
When the object import runs and the JSON contains only an object ID for
a child object, currently nothing happens. This change will create a
new object and set up the relationship with the ID set and the rest of
the attributes to their default values. When future imports happen and
more detail is gathered for these child objects, the stub objects can
then be fleshed out and those relationships are already set.

Without this, you may miss out on that relationship data that was
attempted to be set in the first import, and when the details are added
we don't know how the map goes.