Provide deprecated versions of the shorthand methods
This will aid users making the transition to MR_ prefixed methods, as it will
mark all shorthand uses with a deprecation warning and provide the correct
method to use in the warning. It's pretty nifty.

Shorthand is enabled by default now, but any time it is used there's a warning.

Also, there's a kickass ruby script for regenerating the
MagicalRecordShorthand.[hm] sources automatically. Oh yeah! :boom: :fist bump: